feat(database): track stake pools
