--- database/plugin/metadata/sqlite/models/pool_retirement.go
@@ -1,26 +1,0 @@
-// Copyright 2025 Blink Labs Software
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package models
-
-type PoolRetirement struct {
-	ID          uint   `gorm:"primarykey"`
-	PoolKeyHash []byte `gorm:"index"`
-	Epoch       uint64
-	AddedSlot   uint64
-}
-
-func (PoolRetirement) TableName() string {
-	return "pool_retirement"
-}
